move out Horizon error printing from package
