+++ exp/txnbuild/cmd/txnbuild_examples/main.go
@@ -3,12 +3,14 @@ package main
 // This is a scratch pad for testing new operations. Please DO NOT review!
 
 import (
+	"encoding/json"
 	"log"
 	"os"
 
 	"github.com/stellar/go/clients/horizon"
 	"github.com/stellar/go/exp/txnbuild"
 	"github.com/stellar/go/keypair"
+	"github.com/stellar/go/support/errors"
 )
 
 type key struct {
@@ -46,9 +48,9 @@ func main() {
 	client := horizon.DefaultTestNetClient
 	txnbuild.UseTestNetwork()
 
-	// resp := exampleCreateAccount(client)
+	resp := exampleCreateAccount(client, false)
 	// resp := exampleSendLumens(client, true)
-	resp := exampleBumpSequence(client, true)
+	// resp := exampleBumpSequence(client, true)
 	txnbuild.PrintTransactionSuccess(resp)
 }
 
@@ -102,10 +104,42 @@ func exampleSendLumens(client *horizon.Client, mock bool) horizon.TransactionSuc
 	txeBase64 := buildSignEncode(tx, keys[0].Seed)
 	log.Println("Base 64 TX: ", txeBase64)
 
-	var resp horizon.TransactionSuccess
+	resp := submit(client, txeBase64, mock)
+
+	return resp
+}
+
+func exampleCreateAccount(client *horizon.Client, mock bool) horizon.TransactionSuccess {
+	keys := initKeys()
+	sourceAccount, err := client.LoadAccount(keys[0].Address)
+	dieIfError("loadaccount", err)
+
+	// newAccountKeypair := createKeypair()
+	createAccount := txnbuild.CreateAccount{
+		Destination: "GAS4V4O2B7DW5T7IQRPEEVCRXMDZESKISR7DVIGKZQYYV3OSQ5SH5LVP",
+		Amount:      "10",
+		Asset:       "native",
+	}
+	// inflation := txnbuild.Inflation{}
+
+	tx := txnbuild.Transaction{
+		SourceAccount: sourceAccount,
+		Operations:    []txnbuild.Operation{&createAccount},
+	}
+
+	txeBase64 := buildSignEncode(tx, keys[0].Seed)
+	log.Println("Base 64 TX: ", txeBase64)
+
+	resp := submit(client, txeBase64, mock)
+
+	return resp
+}
+
+func submit(client *horizon.Client, txeBase64 string, mock bool) (resp horizon.TransactionSuccess) {
 	if mock == true {
 		resp = mockSuccess()
 	} else {
+		var err error
 		resp, err = client.SubmitTransaction(txeBase64)
 		if err != nil {
 			bad := err.(*horizon.Error)
@@ -114,46 +148,7 @@ func exampleSendLumens(client *horizon.Client, mock bool) horizon.TransactionSuc
 		}
 	}
 
-	return resp
-}
-
-func exampleCreateAccount(client *horizon.Client) horizon.TransactionSuccess {
-	secretSeed := "SBPQUZ6G4FZNWFHKUWC5BEYWF6R52E3SEP7R3GWYSM2XTKGF5LNTWW4R"
-	sourceAddress := "GDQNY3PBOJOKYZSRMK2S7LHHGWZIUISD4QORETLMXEWXBI7KFZZMKTL3"
-	sourceAccount, err := client.LoadAccount(sourceAddress)
-	dieIfError("loadaccount", err)
-
-	// newAccountKeypair := createKeypair()
-	createAccount := txnbuild.CreateAccount{
-		// Destination: newAccountKeypair.Address(),
-		Destination: "GCCOBXW2XQNUSL467IEILE6MMCNRR66SSVL4YQADUNYYNUVREF3FIV2Z",
-		Amount:      "10",
-		Asset:       "native",
-	}
-	// inflation := txnbuild.Inflation{}
-
-	tx := txnbuild.Transaction{
-		SourceAccount: sourceAccount,
-		// Operations:    []txnbuild.Operation{&inflation},
-		// Operations: []txnbuild.Operation{&inflation, &createAccount},
-		Operations: []txnbuild.Operation{&createAccount},
-	}
-
-	txeBase64 := buildSignEncode(tx, secretSeed)
-	log.Println("Base 64 TX: ", txeBase64)
-
-	// TODO: Add client method to convert to base 64 internally.
-	// resp, err := client.SubmitTransaction(txeBase64)
-	// if err != nil {
-	// 	bad := err.(*horizon.Error)
-	// 	txnbuild.PrintHorizonError(bad)
-	// 	os.Exit(1)
-	// }
-
-	// verify(txeBase64)
-	resp := mockSuccess()
-
-	return resp
+	return
 }
 
 func buildSignEncode(tx txnbuild.Transaction, secretSeed string) string {
@@ -202,4 +197,41 @@ func createKeypair() *keypair.Full {
 	log.Println("Address:", pair.Address())
 
 	return pair
+}
+
+// PrintHorizonError decodes and prints the contents of horizon.Error.Problem.
+// Decoded XDR can be pasted into the Stellar Laboratory XDR viewer
+// (https://www.stellar.org/laboratory) for further analysis.
+// TODO: Move this to new client
+func PrintHorizonError(hError *horizon.Error) error {
+	problem := hError.Problem
+	log.Println("Error type:", problem.Type)
+	log.Println("Error title:", problem.Title)
+	log.Println("Error status:", problem.Status)
+	log.Println("Error detail:", problem.Detail)
+	log.Println("Error instance:", problem.Instance)
+
+	var decodedResultCodes map[string]interface{}
+	var decodedResult, decodedEnvelope string
+	var err error
+
+	err = json.Unmarshal(problem.Extras["result_codes"], &decodedResultCodes)
+	if err != nil {
+		return errors.Wrap(err, "Couldn't unmarshal result_codes")
+	}
+	log.Println("Error extras result codes:", decodedResultCodes)
+
+	err = json.Unmarshal(problem.Extras["result_xdr"], &decodedResult)
+	if err != nil {
+		return errors.Wrap(err, "Couldn't unmarshal result_xdr")
+	}
+	log.Println("Error extras result (TransactionResult) XDR:", decodedResult)
+
+	err = json.Unmarshal(problem.Extras["envelope_xdr"], &decodedEnvelope)
+	if err != nil {
+		return errors.Wrap(err, "Couldn't unmarshal envelope_xdr")
+	}
+	log.Println("Error extras envelope (TransactionEnvelope) XDR:", decodedEnvelope)
+
+	return nil
 }
